Modify authorityKeyIdentifier of the original certificate.
I'm using my personal CA to sign the certificate with the --cert and
--key options, however the certificate generated failed to verify, and
openssl complained about not being able to get local issuer. I did some
inspection and found out that the issuer was indeed correct, but the
authorityKeyIdentifier of it is not changed to the subjectKeyIdentifier
of my CA, I tried to make the change and everything is working fine.

Tested with openssl-1.1.1c.